--- extensions/link/src/stable/createLinkComponent.tsx
@@ -1,12 +1,0 @@
-import type { TypeLink } from "./Link";
-import { Link } from "./Link";
-
-export function createLinkComponent<
-  T extends { [activityName: string]: unknown },
->(): {
-  Link: TypeLink<T>;
-} {
-  return {
-    Link: Link as TypeLink<T>,
-  };
-}
